refactor(inbox): split leaf helpers out of the 3217-line inbox route file
First of three behavior-preserving slices. Moves the pure/leaf helpers out of
packages/dashboard/src/routes/inbox.ts into four cohesive sibling modules:

- inbox-shared.ts  — http/util (publishInboxEvent, addSystemMessage, isAjax,
  parseSort/Flash, latestUserRequest, localIsoNow, parseActionMeta, …) + shared
  constants (SYSTEM_AGENT_IDS, SORT_*, …) + pure formatters (formatLearnings,
  formatConversationSnapshot). Leaf — imports no other inbox-* module.
- inbox-catalog.ts — sub-agent allowlist/catalog/specs + input enrichment
  (getSubAgentAllowlist, ensureSystemAgentCurrent, build*CatalogJson, enrich*Inputs,
  ALLOWLIST_AUTOIMPORT_DIR + its readFileSync site, …).
- inbox-plan.ts    — plan/action/link parsing + crash recovery (parseProposedActions,
  parseTriageLinks, hasMatchingFailedAction, planTriageCrashRecovery, TriageLink, …).
- inbox-widgets.ts — thread view-data + in-thread widget assembly (buildThreadSummary,
  buildInlineActionWidgets, INLINE_INBOX_WIDGET_TYPES, exportTargetAgentYaml, …).

inbox.ts re-exports the moved symbols so the 9 sibling test files keep importing from
./inbox.js untouched. Pure cut-and-paste — no function->const conversions, no logic
changes. inbox.ts: 3217 -> 2249. Clean build; full dashboard+core suite unchanged at
2018 pass / 3 skip. Engine extraction (PR2) + shim cleanup/test repoint (PR3) follow.
